fit-intelligence/requirements.txt
chromadb==1.0.20
flask-cors==6.0.1
numpy==2.3.2
pandas==2.3.2
requests==2.32.5
requests-oauthlib==2.0.0
sentence-transformers==5.1.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
transformers==4.55.3
